Moved Rapid 3DS functions to separate rapid.js file, which is then re-exported in index.js.
