+++ package.json
@@ -7,7 +7,7 @@
   "browser": "dist/index.umd.js",
   "scripts": {
     "build": "rimraf dist && webpack",
-    "doc": "jsdoc2md --template doc/README.hbs --files src/lib/index.js > README.md",
+    "doc": "jsdoc2md --template doc/README.hbs --files src/lib/*.js > README.md",
     "demo": "webpack serve --config webpack.demo.config.js"
   },
   "repository": {
